Fix failing jruby tests, change default REDIS_BRANCH to stable
jruby tests fail when using jruby_openssl 0.10.0 but not when using 0.9.21
This change adds 0.9.21 to the travis build matrix and sets 0.10.0 as an allowed failure.

This change also switches the default REDIS_BRANCH to stable since
the unstable branch of redis isn't always in a buildable state.

Also adds minor tweaks to how the tests are run due to minor issues with Gemfile usage.